requirements.txt
schedule
paho.mqtt
certifi
requests
